// src/components/general/GeneralStyles.types.ts
export interface LayoutProps {
    num?: string | {};
    phoneGap?: string;
    alignFlexStart?: boolean;
    justifyFlexStart?: boolean;
    left?: boolean;
    retainGrid?: boolean;
    tabNum?: string | {};
    gapTab?: string;
    auto?: boolean;
    gap?: string;
    phoneNum?: string | {};
    inline?: string;
    fullHeight?: boolean;
    vertical?: boolean;
    alignFlexEnd?: boolean;
    justifyFlexEnd?: boolean;
    wrapContainer?: boolean;
    spaceBetween?: boolean;
    spaceAround?: boolean;
    verticalReversePhone?: boolean;
    top?: boolean;
    center?: boolean;
}